Corrected some Markdown
To use inline code format takes a single backtick. To format multiple lines, you must enclose them with triple backticks, which must appear at the beginning of the line. In this case the line can have no other text.

So with a single backtick you can `do this` but for multiple lines you must
```
do it
this way
```